Drop appengine/msgpack dependencies
Leans out the package by dropping appengine/msgpack dependencies, made
possible by also dropping the `StructMapKeys` option.

This means that the library can no longer be usable with complex key
types (e.g. where a key is a struct), but that's a reasonable tradeoff
for a far leaner set of dependencies.